backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
boto3==1.34.75
langchain==0.1.12
langchain-community==0.0.29
chromadb==0.4.24
openai-whisper==20240930
whisper==1.1.10
python-dotenv==1.0.1
sentence-transformers==2.6.0
langchain_huggingface==0.0.1
pydantic>=2.5.2 
langchain-text-splitters==0.0.1
openai==1.6.1
tweepy==4.14.0
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
faiss-cpu==1.7.4
numpy>=1.26.0
uuid>=1.30
# DataWhisker x-mcp-server dependencies
aiohttp
mcp
mcp-client
langgraph
python-dotenv
